🚚 api: move pdf read-models to src
