+++ api/src/certification/session/domain/read-models/CertificationCandidateForAttendanceSheet.js
@@ -1,5 +1,4 @@
 import lodash from 'lodash';
-
 const { isNil } = lodash;
 
 class CertificationCandidateForAttendanceSheet {
