+++ api/src/certification/session/domain/read-models/SessionForAttendanceSheet.js
