+++ api/tests/tooling/domain-builder/factory/build-certification-candidate-for-attendance-sheet.js
@@ -1,4 +1,4 @@
-import { CertificationCandidateForAttendanceSheet } from '../../../../lib/domain/read-models/CertificationCandidateForAttendanceSheet.js';
+import { CertificationCandidateForAttendanceSheet } from '../.././../../src/certification/session/domain/read-models/CertificationCandidateForAttendanceSheet.js';
 
 const buildCertificationCandidateForAttendanceSheet = function ({
   lastName = 'Laifrui',
